refactor(client/notes): restructure components by context and improve naming
components/
├── dashboard/
│   └── NotesDashboardPanel.jsx     // Was NotesPanel.jsx
├── layout/
│   ├── NotesPageLayout.jsx         // Was NotesContainer.jsx
│   ├── NoteEditingPanel.jsx        // Was NoteEditor.jsx
│   ├── NotesList.jsx               // Moved from root
│   └── NotesNavbar.jsx             // Moved from root
└── editors/
    ├── DashboardNoteEditor.jsx     // Was MiniTipTapEditor.jsx
    └── RichTextEditor.jsx          // Was TipTapEditor.jsx
