+++ client/src/features/Notes/components/index.js
@@ -1,0 +1,12 @@
+// Dashboard components
+export { default as NotesDashboardPanel } from "./dashboard/NotesDashboardPanel";
+
+// Layout components
+export { default as NotesPageLayout } from "./layout/NotesPageLayout";
+export { default as NoteEditingPanel } from "./layout/NoteEditingPanel";
+export { default as NotesList } from "./layout/NotesList";
+export { default as NotesNavbar } from "./layout/NotesNavbar";
+
+// Editor components
+export { default as DashboardNoteEditor } from "./editors/DashboardNoteEditor";
+export { default as RichTextEditor } from "./editors/RichTextEditor";
